--- a/stack.yaml
+++ b/stack.yaml
@@ -1,10 +1,10 @@
-resolver: lts-8.13
+resolver: lts-9.0
 
 packages:
 - '.'
 
 extra-deps:
-  - hedgehog-0.2
+  - hedgehog-0.5
 
 flags: {}