Implementation of Rank that is nearly three times faster

# Before
```
benchmarking 8-bit/Rank - Once
time                 14.31 ns   (14.17 ns .. 14.45 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 14.36 ns   (14.24 ns .. 14.51 ns)
std dev              481.5 ps   (395.5 ps .. 611.7 ps)
variance introduced by outliers: 55% (severely inflated)

benchmarking 16-bit/Rank - Once
time                 14.82 ns   (14.67 ns .. 15.00 ns)
                     0.999 R²   (0.999 R² .. 0.999 R²)
mean                 14.96 ns   (14.81 ns .. 15.16 ns)
std dev              569.1 ps   (472.1 ps .. 736.7 ps)
variance introduced by outliers: 62% (severely inflated)

benchmarking 32-bit/Rank - Once
time                 15.14 ns   (14.97 ns .. 15.34 ns)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 15.14 ns   (14.97 ns .. 15.36 ns)
std dev              622.6 ps   (505.1 ps .. 798.1 ps)
variance introduced by outliers: 65% (severely inflated)

benchmarking 64-bit/Rank - Once
time                 15.13 ns   (14.97 ns .. 15.30 ns)
                     0.999 R²   (0.999 R² .. 0.999 R²)
mean                 15.32 ns   (15.19 ns .. 15.46 ns)
std dev              475.3 ps   (405.9 ps .. 563.9 ps)
variance introduced by outliers: 51% (severely inflated)
```

# After
```
benchmarking 8-bit/Rank - Once
time                 5.723 ns   (5.670 ns .. 5.788 ns)
                     0.999 R²   (0.999 R² .. 0.999 R²)
mean                 5.761 ns   (5.703 ns .. 5.843 ns)
std dev              224.0 ps   (174.7 ps .. 354.0 ps)
variance introduced by outliers: 64% (severely inflated)

benchmarking 16-bit/Rank - Once
time                 5.849 ns   (5.791 ns .. 5.915 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 5.860 ns   (5.807 ns .. 5.925 ns)
std dev              200.5 ps   (161.5 ps .. 282.4 ps)
variance introduced by outliers: 58% (severely inflated)

benchmarking 32-bit/Rank - Once
time                 5.697 ns   (5.647 ns .. 5.761 ns)
                     0.999 R²   (0.999 R² .. 0.999 R²)
mean                 5.750 ns   (5.692 ns .. 5.816 ns)
std dev              208.6 ps   (168.7 ps .. 270.9 ps)
variance introduced by outliers: 61% (severely inflated)

benchmarking 64-bit/Rank - Once
time                 5.139 ns   (5.089 ns .. 5.193 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 5.144 ns   (5.094 ns .. 5.197 ns)
std dev              180.8 ps   (156.4 ps .. 229.8 ps)
variance introduced by outliers: 59% (severely inflated)
```